requirements.txt
numpy==1.23.5
torch==1.10.0
torchaudio==0.10.0
librosa==0.9.2
pyworld==0.3.2
pysptk==0.2.0
matplotlib==3.6.2
PyYAML==6.0
soundfile==0.11.0
pytorch-lightning==1.9.0
tqdm==4.64.1
pypesq==1.2.4
gradio==2.8.12
fairseq==0.12.2